Добавила интерфейсы для Service + исправила ошибку
