+++ src/main/java/com/smile/blog/repositories/TagRepository.java
@@ -5,4 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 
 public interface TagRepository extends CrudRepository<Tag,Long> {
    Tag findByName(String name);
+   boolean existsByName(String name);
 }
