+++ src/main/java/com/smile/blog/services/TagService.java
@@ -1,43 +1,14 @@
 package com.smile.blog.services;
 
 import com.smile.blog.models.Tag;
-import com.smile.blog.repositories.TagRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
 
-import java.util.ArrayList;
-import java.util.Optional;
+public interface TagService {
 
-@Component
-public class TagService {
+    Iterable<Tag> getAllTag();
 
-    private static TagRepository tagRepository;
+    Tag findTagById(Long id);
 
-    @Autowired
-    public TagService(TagRepository tagRepository) {
-        this.tagRepository = tagRepository;
-    }
+     Tag findTagByName(String tag);
 
-    public static   Iterable<Tag> getAllTag()
-    {
-        return tagRepository.findAll();
-    }
-
-    public static Tag findTagById(Long id){
-        if(!tagRepository.existsById(id)){
-            return null;
-        }
-        Optional<Tag> tags = tagRepository.findById(id);
-        ArrayList<Tag> res = new ArrayList<>();
-        tags.ifPresent(res::add);
-        return res.get(0);
-    }
-
-    public static Tag findTagByName(String tag){
-        return tagRepository.findByName(tag);
-    }
-
-    public static void addTag(String name, String shortDescription){
-        tagRepository.save(new Tag(name,shortDescription));
-    }
+    void addTag(String name, String shortDescription);
 }
